refactor: split backends into a public submodule
Move VectorBackend, GenericBackend, SqliteVecBackend, PgVectorBackend
and the auto_backend dispatcher out of __init__.py and into a new
flask_image_search.backends module.

The top-level package now exports only ImageSearch -- backends are
power-user surface and live behind the submodule import. __init__.py
shrinks from ~830 lines to ~310 and only carries Flask-extension
concerns; backends.py owns its own SQLAlchemy/numpy imports.

Public API change: backend classes are no longer re-exported from
flask_image_search. Callers who pass backend= explicitly switch to
`from flask_image_search.backends import SqliteVecBackend`.
